% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proxies.R
\name{sg_add_node_p}
\alias{sg_add_node_p}
\title{Add node}
\usage{
sg_add_node_p(proxy, nodes, ...)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{nodes}{A \code{data.frame} of nodes.}

\item{...}{any column.}
}
\description{
Proxy to dynamically add a node.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
	actionButton("add node"),
	sigmajsOutput("sg")
)

server <- function(input, output) {
 nodes <- data.frame(
   id = ids,
   label = LETTERS[1:10],
   x = runif(10, 1, 20),
   y = runif(10, 1, 20),
   size = runif(10, 1, 5),
   stringsAsFactors = FALSE
 )

 edges <- data.frame(
   id = 1:15,
   source = sample(ids, 15, replace = TRUE),
   target = sample(ids, 15, replace = TRUE),
   stringsAsFactors = FALSE
 )
 output$sg <- renderSigmajs({
   sigmajs() \%>\%
     sg_nodes(nodes, id, label) \%>\%
     sg_edges(edges, id, source, target)
 })
}

shinyApp(ui, server)
}

}
