% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{sg_add_nodes_delay_p}
\alias{sg_add_nodes_delay_p}
\alias{sg_add_edges_delay_p}
\title{Add nodes or edges with a delay}
\usage{
sg_add_nodes_delay_p(proxy, data, delay, ..., refresh = TRUE)

sg_add_edges_delay_p(proxy, data, delay, ..., refresh = TRUE)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{data}{A \code{data.frame} of _one_ node or edge.}

\item{delay}{Column name containing delay in milliseconds to wait before adding the node or edge.}

\item{...}{any column.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect.}
}
\description{
Proxies to dynamically add multiple nodes or edges to an already existing graph with a *delay* between each addition.
}
\note{
Have the parameters from your initial graph match that of the node you add, i.e.: if you pass \code{size} in your initial chart,
make sure you also have it in your proxy.
}
\examples{
\dontrun{
demo("add-nodes-delay", package = "sigmajs")
demo("add-edges-delay", package = "sigmajs")
demo("add-all-delay", package = "sigmajs")
}

}
