% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{sg_add_nodes_p}
\alias{sg_add_nodes_p}
\alias{sg_add_edges_p}
\title{Add nodes or edges}
\usage{
sg_add_nodes_p(proxy, data, ..., refresh = TRUE, rate = "once")

sg_add_edges_p(proxy, data, ..., refresh = TRUE, rate = "once")
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{data}{A \code{data.frame} of nodes or edges.}

\item{...}{any column.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect, if you are running force the algorithm is killed and restarted at every iteration..}

\item{rate}{Refresh rate, either \code{once}, the graph is refreshed after data.frame of nodes is added or at each \code{iteration} (row-wise). Only applies if \code{refresh} is set to \code{TRUE}.}
}
\value{
The \code{proxy} object.
}
\description{
Proxies to dynamically add *multiple* nodes or edges to an already existing graph.
}
\note{
Have the parameters from your initial graph match that of the node you add, i.e.: if you pass \code{size} in your initial chart,
make sure you also have it in your proxy.
}
\examples{
\dontrun{
demo("add-nodes", package = "sigmajs")
demo("add-edges", package = "sigmajs")
}

}
