% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{sg_animate}
\alias{sg_animate}
\title{Animate}
\usage{
sg_animate(sg, mapping, delay = 5000)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{mapping}{Variables to map animation to.}

\item{delay}{Delay in milliseconds before animation is triggered.}
}
\description{
Animate graph components.
}
\examples{
# generate graph
nodes <- sg_make_nodes(20)
edges <- sg_make_edges(nodes, 30)

# add transition
n <- nrow(nodes)
nodes$to_x <- runif(n, 5, 10)
nodes$to_y <- runif(n, 5, 10)
nodes$to_size <- runif(n, 5, 10)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size, color, to_x, to_y, to_size) \%>\%
  sg_edges(edges, id, source, target) \%>\% 
  sg_animate(mapping = list(x = "to_x", y = "to_y", size = "to_size"))

}
