% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{sg_clear_p}
\alias{sg_clear_p}
\alias{sg_kill_p}
\title{Clear or kill the graph}
\usage{
sg_clear_p(proxy, refresh = TRUE)

sg_kill_p(proxy, refresh = TRUE)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect, if you are running force the algorithm is killed and restarted.}
}
\description{
Clear all nodes and edges from the graph or kills the graph.
}
\examples{
\dontrun{
demo("clear-graph", package = "sigmajs")
}

}
