% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{color-scale}
\alias{color-scale}
\alias{sg_scale_color}
\title{Color}
\usage{
sg_scale_color(sg, pal)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{pal}{Vector of color.}
}
\description{
Scale color by node size.
}
\examples{
nodes <- sg_make_nodes() 
edges <- sg_make_edges(nodes, 20)

sigmajs() \%>\% 
  sg_nodes(nodes, id, size) \%>\% 
  sg_scale_color(pal = c("red", "blue"))

}
