% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{sg_drop_node_p}
\alias{sg_drop_node_p}
\alias{sg_drop_edge_p}
\title{Remove node or edge}
\usage{
sg_drop_node_p(proxy, id, refresh = TRUE)

sg_drop_edge_p(proxy, id, refresh = TRUE)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{id}{Id of edge or node to delete.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect, if you are running force the algorithm is killed and restarted.}
}
\description{
Proxies to dynamically remove a node or an edge to an already existing graph.
}
\examples{
\dontrun{
demo("drop-node", package = "sigmajs")
}

}
