% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{sg_drop_nodes_p}
\alias{sg_drop_nodes_p}
\alias{sg_drop_edges_p}
\title{Drop nodes or edges}
\usage{
sg_drop_nodes_p(proxy, data, ids, refresh = TRUE, rate = "once")

sg_drop_edges_p(proxy, data, ids, refresh = TRUE, rate = "once")
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{data}{A \code{data.frame} of nodes or edges.}

\item{ids}{Column containing ids to drop from the graph.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect.}

\item{rate}{Refresh rate, either \code{once}, the graph is refreshed after data.frame of nodes is added or at each \code{iteration} (row-wise). Only applies if \code{refresh} is set to \code{TRUE}.}
}
\value{
The \code{proxy} object.
}
\description{
Proxies to dynamically drop *multiple* nodes or edges from an already existing graph.
}
\note{
Have the parameters from your initial graph match that of the node you add, i.e.: if you pass \code{size} in your initial chart,
make sure you also have it in your proxy.
}
