% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{sg_filter_gt_p}
\alias{sg_filter_gt_p}
\alias{sg_filter_lt_p}
\title{FIlter}
\usage{
sg_filter_gt_p(proxy, input, var, target = "nodes")

sg_filter_lt_p(proxy, input, var, target = "nodes")
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{input}{A Shiny input.}

\item{var}{Variable to filter.}

\item{target}{Target of filter, \code{nodes}, \code{edges}, or \code{both}.}
}
\description{
Filter nodes and/or edges.
}
\section{Functions}{

\itemize{
  \item{\code{sg_filter_gt_p} Filter greater than \code{var}.}
  \item{\code{sg_filter_lt_p} Filter less than \code{var}.}
}
}

\examples{
# demo("filter-nodes", package = "sigmajs")

}
