% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sg_make_nodes}
\alias{sg_make_nodes}
\alias{sg_make_edges}
\title{Generate data}
\usage{
sg_make_nodes(n = 10, colors = c("#B1E2A3", "#98D3A5", "#328983", "#1C5C70",
  "#24C96B"))

sg_make_edges(nodes, n = 15)
}
\arguments{
\item{n}{Number of nodes.}

\item{colors}{Color palette to use.}

\item{nodes}{Nodes, as generated by \code{sg_make_nodes}.}
}
\value{
\code{tibble} of nodes or edges.
}
\description{
Generate nodes and edges.
}
\examples{
nodes <- sg_make_nodes()
edges <- sg_make_edges(nodes)

sigmajs() \%>\%
sg_nodes(nodes, id, label, size, color) \%>\%
sg_edges(edges, id, source, target) \%>\%
sg_settings(defaultNodeColor = "#0011ff")

}
