% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sg_nodes}
\alias{sg_nodes}
\alias{sg_edges}
\alias{sg_edges2}
\alias{sg_nodes2}
\title{Add nodes and edges}
\usage{
sg_nodes(sg, data, ...)

sg_edges(sg, data, ...)

sg_edges2(sg, data)

sg_nodes2(sg, data)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{data}{Data.frame of nodes or edges.}

\item{...}{any column.}
}
\description{
Add nodes and edges to a \code{sigmajs} graph.
}
\details{
Eaach node must include a unique id, ideally the user passes \code{x} and \code{y}, if they are not passed then they are randomly generated.
Each edge also must include a unique id.
}
\examples{
ids <- as.character(1:10)

nodes <- data.frame(
  id = ids,
  label = LETTERS[1:10],
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:15,
  source = sample(ids, 15, replace = TRUE),
  target = sample(ids, 15, replace = TRUE),
  type = "curvedArrow",
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label) \%>\%
  sg_edges(edges, id, source, target)

}
