% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesmis_edges}
\alias{lesmis_edges}
\title{Edges from co-appearances of characters in "Les Miserables"}
\format{An igraph object with 181 nodes and 4 variables
\describe{
  \item{\code{source}}{ abbreviation of character name}
  \item{\code{target}}{ abbreviation of character name}
  \item{\code{id}}{ unique edge id}
 \item{\code{label}}{ edge label}
}}
\source{
\url{https://github.com/MADStudioNU/lesmiserables-character-network}
}
\usage{
lesmis_edges
}
\description{
A graph where the nodes are characters in "Les Miserables" updated from its first encoding 
by Professor Donald Knuth, as part of the Stanford Graph Base (SGB)
}
\keyword{datasets}
