% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesmis_igraph}
\alias{lesmis_igraph}
\title{Co-appearances of characters in "Les Miserables" as igraph object}
\format{An igraph object with 181 nodes and 1589 edges
\describe{
  \item{\code{id}}{ abbreviation of character name}
  \item{\code{label}}{ character name}
  \item{\code{color}}{ random color}
}}
\source{
\url{https://github.com/MADStudioNU/lesmiserables-character-network}
}
\usage{
lesmis_igraph
}
\description{
A graph where the nodes are characters in "Les Miserables" updated from its first encoding 
by Professor Donald Knuth, as part of the Stanford Graph Base (SGB)
}
\keyword{datasets}
