% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesmis_nodes}
\alias{lesmis_nodes}
\title{Nodes from co-appearances of characters in "Les Miserables"}
\format{An igraph object with 181 nodes and 2 variables
\describe{
  \item{\code{id}}{ abbreviation of character name}
  \item{\code{label}}{ character name}
}}
\source{
\url{https://github.com/MADStudioNU/lesmiserables-character-network}
}
\usage{
lesmis_nodes
}
\description{
A graph where the nodes are characters in "Les Miserables" updated from its first encoding 
by Professor Donald Knuth, as part of the Stanford Graph Base (SGB)
}
\keyword{datasets}
