% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{sg_neighbours}
\alias{sg_neighbours}
\alias{sg_neighbors}
\title{Neighbours}
\usage{
sg_neighbours(sg)

sg_neighbors(sg)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}
}
\description{
Highlight node neighbours on click.
}
\examples{
nodes <- sg_make_nodes() 
edges <- sg_make_edges(nodes, 17)

sigmajs() \%>\% 
  sg_nodes(nodes, id, size, color) \%>\% 
  sg_edges(edges, id, source, target) \%>\% 
  sg_layout() \%>\% 
  sg_neighbours()

}
