% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R
\name{read-batch}
\alias{read-batch}
\alias{sg_read_batch_nodes_p}
\alias{sg_read_batch_edges_p}
\alias{sg_read_batch_exec_p}
\title{Batch read}
\usage{
sg_read_batch_nodes_p(proxy, data, ..., batch)

sg_read_batch_edges_p(proxy, data, ..., batch)

sg_read_batch_exec_p(proxy)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{data}{A \code{data.frame} of _one_ node or edge.}

\item{...}{any column.}

\item{batch}{Column name of containing batch identifier.}
}
\description{
Read nodes and edges by batch.
}
\examples{
library(shiny)

ui <- fluidPage(
	actionButton("add", "add nodes & edges"),
	sigmajsOutput("sg")
)

server <- function(input, output, session){

	output$sg <- renderSigmajs({
		sigmajs()
	})

	observeEvent(input$add, {
		nodes <- sg_make_nodes(50)
		nodes$batch <- c(
			rep(1, 25),
			rep(2, 25)
		)

		edges <- data.frame(
			id = 1:80,
			source = c(
				sample(1:25, 40, replace = TRUE),
				sample(26:50, 40, replace = TRUE)
			),
			target = c(
				sample(1:25, 40, replace = TRUE),
				sample(26:50, 40, replace = TRUE)
			),
			batch = c(
				rep(1, 40),
				rep(2, 40)
			)
		) \%>\% 
		dplyr::mutate_all(as.character)

		sigmajsProxy("sg") \%>\% 
			sg_read_batch_nodes_p(nodes, id, color, label, size, batch = "batch") \%>\% 
			sg_read_batch_edges_p(edges, id, source, target, batch = "batch") \%>\% 
			sg_read_batch_exec_p() \%>\% 
			sg_refresh_p()
	})

}

if(interactive()) shinyApp(ui, server)

}
