% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-shapes.R
\name{sg_add_images}
\alias{sg_add_images}
\title{Add images to nodes}
\usage{
sg_add_images(sg, data, url, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{data}{Data.frame containing columns.}

\item{url}{URL of image.}

\item{...}{Any other column.}
}
\description{
Add images to nodes with the \href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.renderers.customShapes}{Custom Shapes plugin}.
}
\examples{
\dontrun{
demo("custom-shapes", package = "sigmajs")
}

}
\seealso{
\href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.renderers.customShapes}{Official documentation}
}
