% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-shapes.R
\name{sg_custom_shapes}
\alias{sg_custom_shapes}
\title{Custom shapes}
\usage{
sg_custom_shapes(sg)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}
}
\description{
Indicate a graph uses custom shapes
}
