% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{sg_force}
\alias{sg_force}
\title{Add force}
\usage{
sg_force(sg, worker = TRUE, barnes.hut = FALSE)
}
\description{
Add force
}
\examples{
ids <- as.character(1:26)

nodes <- data.frame(
  id = ids,
  label = LETTERS,
  size = runif(26, 1, 5),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:40,
  source = sample(ids, 40, replace = TRUE),
  target = sample(ids, 40, replace = TRUE),
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label) \%>\%
  sg_edges(edges, id, source, target) \%>\%
  sg_force()

}
