% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_gexf.R
\name{sg_from_gexf}
\alias{sg_from_gexf}
\title{Graph from GEXF file}
\usage{
sg_from_gexf(sg, file, sd = NULL)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{file}{Path to GEXF file.}

\item{sd}{A \link[crosstalk]{SharedData} of nodes.}
}
\description{
Create a sigmajs graph from a GEXF file.
}
\examples{
gexf <- system.file("examples/arctic.gexf", package = "sigmajs")

sigmajs() \%>\% 
  sg_from_gexf(gexf) 

}
