% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{sg_from_igraph}
\alias{sg_from_igraph}
\title{Create from igraph}
\usage{
sg_from_igraph(sg, igraph, layout = NULL)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{igraph}{An object of class \code{igraph}}

\item{layout}{A matrix of coordinates.}
}
\description{
Create a \code{sigmajs} from an \code{igraph} object.
}
\examples{
\dontrun{
data("lesmis_igraph")

sigmajs() \%>\%
	sg_from_igraph(lesmis_igraph)
}

}
