% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sg_kill}
\alias{sg_kill}
\title{Kill}
\usage{
sg_kill(sg)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}
}
\description{
Kill the graph to ensure new data is redrawn, useful in Shiny 
when graph is not updated bz \code{\link{sigmajsProxy}}.
}
