% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{sg_layout}
\alias{sg_layout}
\title{Layouts}
\usage{
sg_layout(sg, directed = TRUE, layout = igraph::layout_nicely, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{directed}{Whether or not to create a directed graph, passed to \code{\link[igraph]{graph_from_data_frame}}.}

\item{layout}{An \code{igraph} layout function.}

\item{...}{Any other parameter to pass to \code{layout} function.}
}
\description{
Layout your graph.
}
\examples{
nodes <- sg_make_nodes(250) # 250 nodes
edges <- sg_make_edges(nodes, n = 500)

sigmajs() \%>\% 
  sg_nodes(nodes, id, size, color) \%>\% 
  sg_edges(edges, id, source, target) \%>\% 
  sg_layout()

}
