% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{sg_refresh_p}
\alias{sg_refresh_p}
\title{Refresh instance}
\usage{
sg_refresh_p(proxy)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}
}
\description{
Refresh your instance.
}
\details{
It is often required to refresh the instance when using proxies.
}
