% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-size.R
\name{sg_relative_size}
\alias{sg_relative_size}
\title{Relative node sizes}
\usage{
sg_relative_size(sg, initial = 1)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{initial}{Initial node size.}
}
\value{
A modified version of the \code{sg} object.
}
\description{
Change nodes size depending to their degree (number of relationships)
}
\examples{
nodes <- sg_make_nodes(50)
edges <- sg_make_edges(nodes, 100)

sigmajs() \%>\%
  sg_nodes(nodes, id, label) \%>\% # no need to pass size
  sg_edges(edges, id, source, target) \%>\% 
  sg_relative_size()

}
