% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{sg_settings}
\alias{sg_settings}
\title{Settings}
\usage{
sg_settings(sg, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{...}{Any parameter, see \href{https://github.com/jacomyal/sigma.js/wiki/Settings}{official documentation}.}
}
\description{
Graph settings.
}
\examples{
ids <- as.character(1:10)

nodes <- data.frame(
  id = ids,
  label = LETTERS[1:10],
  x = runif(10, 1, 20),
  y = runif(10, 1, 20),
  size = runif(10, 1, 5),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:15,
  source = sample(ids, 15, replace = TRUE),
  target = sample(ids, 15, replace = TRUE),
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size, x, y) \%>\%
  sg_edges(edges, id, source, target) \%>\% 
  sg_force() \%>\% 
  sg_settings(
    defaultNodeColor = "#0011ff"
  )

}
