% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{sg_zoom_p}
\alias{sg_zoom_p}
\title{Zoom}
\usage{
sg_zoom_p(proxy, id, ratio = 0.5, duration = 1000)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{id}{Node id to zoom to.}

\item{ratio}{The zoom ratio of the graph and its items.}

\item{duration}{Duration of animation.}
}
\description{
Dynamically Zoom a node.
}
