% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proxies.R
\name{sg_add_node_p}
\alias{sg_add_node_p}
\alias{sg_add_edge_p}
\title{Add node or edge}
\usage{
sg_add_node_p(proxy, node, ...)

sg_add_edge_p(proxy, edge, ...)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{node, edge}{A \code{data.frame} of _one_ node or edge.}

\item{...}{any column.}
}
\description{
Proxy to dynamically add a node or an edge to already existing graph.
}
\note{
Have the parameters from your initial graph match that of the node you add, i.e.: if you pass \code{size} in your initial chart,
make sure you also have it in your proxy.
}
\examples{
\dontrun{
demo("add-node", package = "sigmajs")
demo("add-edge", package = "sigmajs")
}

}
