% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_proxies.R, R/helpers.R
\name{sg_clear_p}
\alias{sg_clear_p}
\alias{sg_kill_p}
\alias{sg_kill}
\alias{sg_clear}
\title{Clear or kill the graph}
\usage{
sg_clear_p(proxy, refresh = TRUE)

sg_kill_p(proxy, refresh = TRUE)

sg_kill(sg)

sg_clear(sg)
}
\arguments{
\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect, if you are running force the algorithm is killed and restarted.}

\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}
}
\value{
The \code{proxy} object.

A modified version of the \code{sg} object.
}
\description{
Clear all nodes and edges from the graph or kills the graph.

Kill the graph to ensure new data is redrawn, useful in Shiny 
when graph is not updated by \code{\link{sigmajsProxy}}.
}
