% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{sg_force}
\alias{sg_force}
\alias{sg_force_restart_p}
\alias{sg_force_start_p}
\alias{sg_force_stop_p}
\alias{sg_force_config_p}
\title{Add force}
\usage{
sg_force(sg, ...)

sg_force_restart_p(proxy, ...)

sg_force_start_p(proxy, ...)

sg_force_stop_p(proxy)

sg_force_config_p(proxy, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{...}{Any parameter, see \href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.layout.forceAtlas2}{official documentation}.}

\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}
}
\description{
Implementation of \href{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0098679}{forceAtlas2}.
}
\examples{
ids <- as.character(1:10)

nodes <- data.frame(
  id = ids,
  label = LETTERS[1:10],
  size = runif(10, 1, 5),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:15,
  source = sample(ids, 15, replace = TRUE),
  target = sample(ids, 15, replace = TRUE),
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size) \%>\%
  sg_edges(edges, id, source, target) \%>\% 
  sg_force()

}
