% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{sg_force}
\alias{sg_force}
\alias{sg_force_stop}
\alias{sg_force_restart_p}
\alias{sg_force_start_p}
\alias{sg_force_stop_p}
\alias{sg_force_kill_p}
\alias{sg_force_config_p}
\title{Add force}
\usage{
sg_force(sg, ...)

sg_force_stop(sg, delay = 5000)

sg_force_restart_p(proxy, ..., refresh = TRUE)

sg_force_start_p(proxy, ..., refresh = TRUE)

sg_force_stop_p(proxy)

sg_force_kill_p(proxy)

sg_force_config_p(proxy, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{...}{Any parameter, see \href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.layout.forceAtlas2}{official documentation}.}

\item{delay}{Milliseconds after which the layout algorithm should stop running.}

\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}

\item{refresh}{Whether to refresh the graph after node is dropped, required to take effect.}
}
\description{
Implementation of \href{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0098679}{forceAtlas2}.
}
\section{Functions}{

\itemize{
\item{\code{sg_force} starts the forceAtlas2 layout}
\item{\code{sg_force_stop} stops the forceAtlas2 layout after a \code{delay} milliseconds}
\item{\code{sg_force_restart_p} proxy to re-starts (\code{kill} then \code{start}) the forceAtlas2 layout, the options you pass to this function are applied on restart. If forceAtlas2 has not started yet it is launched.}
\item{\code{sg_force_start_p} proxy to start forceAtlas2.}
\item{\code{sg_force_stop_p} proxy to stop forceAtlas2.}
\item{\code{sg_force_kill_p} proxy to ompletely stops the layout and terminates the assiociated worker. You can still restart it later, but a new worker will have to initialize.}
\item{\code{sg_force_config_p} proxy to set configurations of forceAtlas2.}
}
}

\examples{
nodes <- sg_make_nodes(50)
edges <- sg_make_edges(nodes, 100)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size) \%>\%
  sg_edges(edges, id, source, target) \%>\% 
  sg_force() \%>\% 
	 sg_force_stop() # stop force after 5 seconds

}
\seealso{
\href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.layout.noverlap}{official documentation}
}
