% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sg_nodes}
\alias{sg_nodes}
\alias{sg_edges}
\title{Add nodes}
\usage{
sg_nodes(sg, nodes, ...)

sg_edges(sg, edges, ...)
}
\arguments{
\item{nodes}{Data.frame of nodes.}

\item{...}{any column.}
}
\description{
Add nodes.
}
\examples{
nodes <- data.frame(
  id = 1:10,
  name = LETTERS[1:10]
)

edges <- data.frame(
  source = sample(1:10, 15, replace = TRUE),
  target = sample(1:10, 15, replace = TRUE)
)

sigmajs() \%>\%
  sg_nodes(nodes, "id", "name") \%>\%
  sg_edges(edges, "source", "target")

}
