% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sg_nodes}
\alias{sg_nodes}
\alias{sg_edges}
\title{Add nodes}
\usage{
sg_nodes(sg, nodes, ...)

sg_edges(sg, edges, ...)
}
\arguments{
\item{nodes}{Data.frame of nodes.}

\item{...}{any column.}
}
\description{
Add nodes.
}
\examples{
ids <- as.character(1:10)

nodes <- data.frame(
  id = ids,
  label = LETTERS[1:10],
  x = runif(10, 1, 20),
  y = runif(10, 1, 20),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:15,
  source = sample(ids, 15, replace = TRUE),
  target = sample(ids, 15, replace = TRUE),
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, x, y) \%>\%
  sg_edges(edges, id, source, target)

}
