% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noverlap.R
\name{sg_noverlap}
\alias{sg_noverlap}
\alias{sg_noverlap_p}
\title{No overlap}
\usage{
sg_noverlap(sg, start = TRUE, ...)

sg_noverlap_p(proxy)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{start}{Whether to start running the noverlap layout.}

\item{...}{any option to pass to the plugin, see \href{https://github.com/jacomyal/sigma.js/tree/master/plugins/sigma.layout.noverlap}{official documentation}.}

\item{proxy}{An object of class \code{sigmajsProxy} as returned by \code{\link{sigmajsProxy}}.}
}
\description{
This plugin runs an algorithm which distributes nodes in the network, ensuring that they do not overlap and providing a margin where specified.
}
\examples{
ids <- as.character(1:10)

nodes <- data.frame(
  id = ids,
  label = LETTERS[1:10],
  size = runif(10, 1, 5),
  stringsAsFactors = FALSE
)

edges <- data.frame(
  id = 1:15,
  source = sample(ids, 15, replace = TRUE),
  target = sample(ids, 15, replace = TRUE),
  stringsAsFactors = FALSE
)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size) \%>\%
  sg_edges(edges, id, source, target) \%>\%
  sg_settings(defaultNodeColor = "#0011ff") \%>\%
  sg_force() \%>\%
  sg_noverlap()

}
