% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{sg_settings}
\alias{sg_settings}
\title{Settings}
\usage{
sg_settings(sg, ...)
}
\arguments{
\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}

\item{...}{Any parameter, see \href{https://github.com/jacomyal/sigma.js/wiki/Settings}{official documentation}.}
}
\description{
Graph settings.
}
\examples{
nodes <- sg_make_nodes()

edges <- sg_make_edges(nodes, 50)

sigmajs() \%>\%
  sg_nodes(nodes, id, label, size) \%>\%
  sg_edges(edges, id, source, target) \%>\% 
  sg_force() \%>\% 
  sg_settings(
    defaultNodeColor = "#0011ff"
  )

}
