% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmajs.R
\name{sigmajs}
\alias{sigmajs}
\title{Initialise}
\usage{
sigmajs(width = "100\%", height = NULL, elementId = NULL)
}
\arguments{
\item{width, height}{Dimensions of graph.}

\item{elementId}{Id of elment.}
}
\description{
Initialise a graph.
}
\note{
Keep \code{width} at \code{100\%} for a responsive visualisation.
}
